move `get_wallet_class` to wallets module
